refactor : auction detail 관련 체리픽
